% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ygpdf_folder.R
\name{get_current_folder}
\alias{get_current_folder}
\title{A ygpdf Function}
\usage{
get_current_folder()
}
\description{
Gives the current folder
}
\examples{
current_dir <- get_current_folder()
current_dir
[1] "D:/github.com/yannick-gouedard/ygpdf/r/ygpdf"
Warning message:
  In get_current_folder() : current_dir from getwd()

NB: A warning message appeared as I directly called the function,
without being in a script

}
\keyword{Current}
\keyword{folder}
