% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ygpdf_process.R
\name{get_r_ygpdf_root_folder}
\alias{get_r_ygpdf_root_folder}
\title{A ygpdf Function}
\usage{
get_r_ygpdf_root_folder()
}
\description{
Gives the R ygpdf root folder
}
\examples{

}
\keyword{R}
\keyword{folder}
\keyword{root}
\keyword{ygpdf}
