% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ygpdf_folder.R
\name{set_GITROOT_ygpdf_folder}
\alias{set_GITROOT_ygpdf_folder}
\title{A ygpdf Function}
\usage{
set_GITROOT_ygpdf_folder()
}
\description{
sets the R ygpdf root folder
}
\examples{

}
\keyword{R}
\keyword{folder}
\keyword{root}
\keyword{ygpdf}
