% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_pdf.R
\name{stamp_pdf}
\alias{stamp_pdf}
\title{A ygpdf Function}
\usage{
stamp_pdf(in_file, this_list = NULL, out_file = NULL)
}
\description{
stamps an existing pdf with an image or a text
}
\examples{

stamp_file <- 'logo_conspel.png'
original_pdf_no_spaces <- gsub(' ', '', original_pdf)           # I do not want any spaces in pdf name
                                                                # so beware if original_pdf has got a path
stamp_list <- data.frame(what = c(stamp_file, stamp_file),      # what?
                         page = c(1, 2),                        # on wich page?
                         type = c('image', 'image'),            # image or text?
                         position_x = c(0.1, 0.5),              # ratio of page width
                         position_y = c(0.2, 0.5),              # ratio of page height
                         size = c(0.1, 0.5),                    # ratio of page width
                         angle = c(0, 0),
                         stringsAsFactors = FALSE)
stamp_pdf(original_pdf_no_spaces, stamp_list, NULL)

Note that:   if (is.null(out_file)) { out_file <- gsub('.pdf', '-stamped.pdf', in_file) }

}
\keyword{adding}
\keyword{image,}
\keyword{pdf,}
\keyword{text}
