% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ygpdf_diff.R
\name{ygpdf_diff}
\alias{ygpdf_diff}
\title{A ygpdf Function}
\usage{
ygpdf_diff(
  first_pdf,
  second_pdf,
  density = 72,
  combine_results = TRUE,
  is_landscape_orientation = TRUE
)
}
\description{
shows differences between 2 pdfs
uses code from https://cran.r-project.org/web/packages/imager/vignettes/gettingstarted.html
}
\examples{

ygpdf_diff(first_pdf, second_pdf, density, combine_results = TRUE, is_landscape_orientation = TRUE)

[ TODO ]: manage portrait / landscape orientation

}
\keyword{differences}
\keyword{pdf,}
