% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ygpdf_stamp.R
\name{ygpdf_stamp}
\alias{ygpdf_stamp}
\title{A ygpdf Function}
\usage{
ygpdf_stamp(in_file, in_list = NULL, out_file = NULL, is_landscape = TRUE)
}
\description{
stamps an existing pdf with an image or a text
}
\examples{

stamp_file <- system.file("images", "Rlogo.pdf", package = "ygpdf")

stamp_list <- data.frame(what = c(stamp_file, stamp_file),  # what?
                         page = c(1, 2),                    # on wich page?
                         type = c('image', 'image'),        # image or text?
                         position_x = c(0.1, 0.5),          # ratio of page width
                         position_y = c(0.2, 0.5),          # ratio of page height
                         size = c(0.1, 0.5),                # ratio of page width
                         angle = c(0, 0),                   # rotation angle
                         alpha = c(0.5, 0.5),               # transparency
                         stringsAsFactors = FALSE)
ygpdf_stamp(original_pdf, stamp_list, NULL, is_landscape = TRUE)

Note that:   if (is.null(out_file)) { out_file <- gsub('.pdf', ' - stamped.pdf', in_file) }

}
\keyword{adding}
\keyword{image,}
\keyword{pdf,}
\keyword{text}
